// src/generated/prisma/models/Category.ts
export type CategoryModel = runtime.Types.Result.DefaultSelection<Prisma.$CategoryPayload>
// ...
export type $CategoryPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Category"
  objects: {
    products: Prisma.$ProductPayload<ExtArgs>[]
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    name: string
    slug: string
    description: string | null
    isActive: boolean
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["category"]>
  composites: {}
}